Added "random_fortune" view.  Updated validate_doc_update.  Updated human readable design doc.
Changed fortune document structure, and made changes in code to reflect
that.

Instead of a sequence_id, our fortune documents will now how a
"random_id".  This random_id is a random number in [0, 1) that will be
used construct a view.  That view will be used to then select a random
document for displaying a random fortune in the web page.